scripts/monitoring/requirements.txt
redis
psycopg2-binary
requests
streamlit
plotly
pandas
pytest
